+++ .python_history
@@ -1,0 +1,4 @@
+import Contest
+import inspect
+inspect.getmembers(Contest, predicate=inspect.ismethod)
+inspect.getmembers(Contest, predicate=inspect.isfunction)
