+++ battle-dj.py
@@ -1,0 +1,33 @@
+from Contest import Fighter
+from Contest import Monster
+
+# In Contest 
+
+lev = Fighter("Lev", 3, 4, 4)
+eli = Fighter("Eli", 5, 3, 4)
+dom = Fighter("Dom", 5, 2, 4)
+
+hurley = Fighter("Hurley", 5, 5, 5)
+
+Getoutman = Monster("Get out man",50,50,50)
+
+
+  
+# eli.battle(lev)
+
+# nitro.sheet()
+lev.sheet()   
+
+
+# Right now we can make Fighter and Monster
+# We can sheet, attack, heal, 
+
+
+# Other ideas 
+# - better stat sheets?
+# - could fighters have quirks?
+# - different weapons?
+# - potions?
+# - monster fights?
+# - faster character creation?
+# - I think dexterity is too important right now ... we need to figure out how to make it less powerful! - DJ
